:truck: move CertificationContract to services
