+++ api/src/certification/evaluation/domain/services/scoring/scoring-v2.js
@@ -22,8 +22,8 @@ import {
 import { AssessmentResultFactory } from '../../../../scoring/domain/models/factories/AssessmentResultFactory.js';
 import { AlgorithmEngineVersion } from '../../../../shared/domain/models/AlgorithmEngineVersion.js';
 import { CertifiedScore } from '../../../domain/models/CertifiedScore.js';
-import { CertificationContract } from '../../models/CertificationContract.js';
 import { CertifiedLevel } from '../../models/CertifiedLevel.js';
+import { CertificationContract } from '../CertificationContract.js';
 
 /**
  * @param {Object} params
